+++ packages/leva/src/types/public.ts
@@ -153,23 +153,23 @@ type PrimitiveToValue<S> = S extends CustomInput<infer I>
   ? boolean
   : NotAPrimitiveType
 
-export type SchemaToValues<S> = BeautifyUnionType<UnionToIntersection<Leaves<S>>>
+export type SchemaToValues<Schema, IncludeTransient extends boolean = false> = BeautifyUnionType<
+  UnionToIntersection<Leaves<IncludeTransient, Schema>>
+>
 
 type Leaf = { ___leaf: 'leaf' }
 type Join<T, K extends keyof T, P> = Leaf extends P ? { [i in K]: T[K] } : P
 
-export type Leaves<T, P extends string | number | symbol = ''> = {
+export type Leaves<IncludeTransient extends boolean, T, P extends string | number | symbol = ''> = {
   // if it's a folder we run the type check on it's schema key
   0: T extends { schema: infer F } ? { [K in keyof F]: Join<F, K, F[K]> } : never
   1: never
   // if the leaf is an object, we run the type check on each of its keys
-  2: T extends { onChange: any } // if an input has the onChange property then it's transient and isn't returned
-    ? never
-    : {
-        [i in P]: T extends { optional: true } ? PrimitiveToValue<T> | undefined : PrimitiveToValue<T>
-      }
+  2: {
+    [i in P]: T extends { optional: true } ? PrimitiveToValue<T> | undefined : PrimitiveToValue<T>
+  }
   // recursivity
-  3: { [K in keyof T]: Join<T, K, Leaves<T[K], K>> }[keyof T]
+  3: { [K in keyof T]: Join<T, K, Leaves<IncludeTransient, T[K], K>> }[keyof T]
   // dead end
   4: Leaf
 }[P extends ''
@@ -182,6 +182,10 @@ export type Leaves<T, P extends string | number | symbol = ''> = {
   ? T extends object
     ? 3
     : 4
+  : T extends { onChange: any } // if an input has the onChange property then it's transient and isn't returned
+  ? IncludeTransient extends true
+    ? 2
+    : 1
   : 2]
 
 /**