+++ packages/leva/CHANGELOG.md
@@ -1,5 +1,17 @@
 # leva
 
+## 0.7.2
+
+### Patch Changes
+
+- d2eaf58: fix: properly resolves value-keyed objects for the `Select` input.
+
+## 0.7.1
+
+### Patch Changes
+
+- 98984e1: types: fix `set` types in transient mode.
+
 ## 0.7.0
 
 ### Minor Changes